[expectations] Handle Regexp like String when apportioning args
If the first argument is a Regexp, it's a message _matcher_, rather than
a message.. but certainly it's not an exception! I'm actually a bit
surprised this doesn't cause any other problems, but it doesn't seem to.

---
This commit was imported from https://github.com/rspec/rspec-expectations/commit/ee27f826663aa524fb66448ec584a8f2d4a77ab1.