[core] Use stronger path-matching for simplecov filters
The filter for spec/ was actually matching basically everything in the
gem (and discarding it), because it's all in lib/rspec/, and simplecov
is doing a string-based infix match. I'm not certain when that changed,
but my guess is 2018, in this PR:

    https://github.com/simplecov-ruby/simplecov/pull/616

I also adjusted the minimum_coverage down to 99, since we're currently
at 99.26, and not 100%. I'll get on that shortly.

---
This commit was imported from https://github.com/rspec/rspec-core/commit/8e7ab794ef40e57bcbcfba4989329ef068885d70.