[core] Add a spec covering the Bisect::ShellRunner.name method
This is a pretty silly spec, but the name method here isn't actually
used by anything that I can see - I assume it exists only for symmetry
with the ShellRunner, to present a more consistent interface). Happy
to just drop the method if that's preferred, and if someone else is
confident that it's not relevant to anything.

---
This commit was imported from https://github.com/rspec/rspec-core/commit/03689c208fd73e72df379b48a8f90d92c9fb78db.