[core] In `RSpec::Core::Bisect::ExampleMinimizer#abort_if_ordering_inconsistent`, minimizer checks the execution order of the devided ids are the same as the original order. So we need to sort them as the original order before run.
For example,
When we have 4 examples: 1.rb, 2.rb, 3.rb, 4.rb and 3.rb is failure order-dependent with 2.rb.
We expect to minimizer reproduce command `rspec 2.rb[1] 3.rb[1]`, but minimizer raises 'The example ordering is inconsistent.'
This is because the ids to run are in the order 4.rb 2.rb (diffrent from the original), so fixed to sort them in the same order as the original error.

---
This commit was imported from https://github.com/rspec/rspec-core/commit/a7219fe7b617a330cb6b59c3d7fe34db757aa70d.